+++ src/screens/account/Agreement.js
